# src/main/resources/static/openapi_v1.yaml
openapi: 3.0.3
info:
  description: 'API para o Tech Challenge da Fase 2 - Pós Tech.


    Este serviço gerencia o cadastro usuários, tipos de usuários do sistema, restaurantes e  seus cardápios.'
  title: ChefManagerApi
  version: 2.0.0
paths:
  /itens-cardapio/{idRestaurante}:
    get:
      parameters:
      - in: path
        name: idRestaurante
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ItemCardapioResponse'
                type: array
          description: Lista retornada
      summary: Lista itens de cardápio
      tags:
      - ItemCardapio
    post:
      parameters:
      - in: path
        name: idRestaurante
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCardapioRequest'
        required: true
      responses:
        '201':
          description: Item criado
      summary: Cria item de cardápio
      tags:
      - ItemCardapio
  /itens-cardapio/{id}:
    delete:
      parameters:
      - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Excluído
      summary: Remove item do cardápio
      tags:
      - ItemCardapio
    get:
      parameters:
      - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemCardapioResponse'
          description: Item encontrado
      summary: Consulta item por ID
      tags:
      - ItemCardapio
    put:
      parameters:
      - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCardapioRequest'
        required: true
      responses:
        '200':
          description: Atualizado
      summary: Atualiza item
      tags:
      - ItemCardapio
  /restaurantes:
    get:
      operationId: buscarRestaurantePorNome
      parameters:
        - description: Nome do restaurante a ser buscado
          in: query
          name: nome
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RestauranteResponse'
                type: array
          description: Lista retornada
        '404':
          description: Restaurante não encontrado com o nome fornecido.
      summary: Lista restaurantes
      tags:
      - Restaurante
    post:
      operationId: criarRestaurante
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestauranteRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestauranteResponse'
          description: Restaurante criado
      summary: Cadastra restaurante
      tags:
      - Restaurante
  /restaurantes/{id}:
    delete:
      operationId: deletarRestaurantePorId
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Restaurante deletado com sucesso
      summary: Deleta restaurante por ID
      tags:
      - Restaurante
    get:
      operationId: buscarRestaurantePorId
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: uuid
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestauranteResponse'
          description: Restaurante encontrado
      summary: Consulta restaurante por ID
      tags:
      - Restaurante
    put:
      operationId: atualizarRestaurante
      parameters:
      - description: ID único do usuário (UUID)
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizarRestauranteRequest'
        description: Dados do restaurante para atualização.
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestauranteResponse'
          description: Usuário atualizado com sucesso.
        '404':
          description: Restaurante não encontrado.
      summary: Atualiza restaurante
      tags:
      - Restaurante
  /tipos-usuario:
    get:
      description: Retorna todos os tipos de usuário cadastrados.
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TipoUsuarioResponse'
                type: array
          description: Lista de tipos retornada com sucesso
      summary: Lista os tipos de usuário
      tags:
      - TipoUsuario
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TipoUsuarioRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TipoUsuarioResponse'
          description: Tipo criado com sucesso
      summary: Cadastra novo tipo de usuário
      tags:
      - TipoUsuario
  /tipos-usuario/{id}:
    delete:
      parameters:
      - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Removido com sucesso
      summary: Remove tipo de usuário
      tags:
      - TipoUsuario
    get:
      parameters:
      - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TipoUsuarioResponse'
          description: Tipo encontrado
      summary: Consulta tipo de usuário por ID
      tags:
      - TipoUsuario
    put:
      parameters:
      - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TipoUsuarioRequest'
        required: true
      responses:
        '200':
          description: Atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TipoUsuarioResponse'
      summary: Atualiza tipo de usuário
      tags:
      - TipoUsuario
  /usuarios:
    get:
      operationId: buscarPorNome
      parameters:
      - description: Nome do usuário a ser buscado
        in: query
        name: nome
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
          description: Usuário encontrado
        '404':
          description: Usuário não encontrado com o nome fornecido.
      summary: Busca um usuário por nome
      tags:
      - Usuario
    post:
      operationId: criarUsuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
          description: Usuário criado com sucesso
      summary: Cadastra um novo usuário
      tags:
      - Usuario
  /usuarios/{id}:
    delete:
      operationId: deletarUsuarioPorId
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '204':
          description: Usuário deletado com sucesso
      summary: Deleta usuário por ID
      tags:
      - Usuario
    get:
      operationId: buscarUsuarioPorId
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
          description: Usuário encontrado
      summary: Busca usuário por ID
      tags:
      - Usuario
    put:
      operationId: atualizar
      parameters:
      - description: ID único do usuário (UUID)
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizarUsuarioRequest'
        description: Dados do usuário para atualização.
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
          description: Usuário atualizado com sucesso.
        '404':
          description: Usuário não encontrado.
      summary: Atualiza um usuário existente
      tags:
      - Usuario
  /usuarios/{id}/senha:
    patch:
      operationId: alterarSenha
      parameters:
      - description: ID do usuário (UUID)
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlterarSenhaRequest'
        description: Dados para alteração de senha.
        required: true
      responses:
        '204':
          description: Senha alterada com sucesso
        '400':
          description: Dados inválidos ou senha incorreta
        '404':
          description: Usuário não encontrado
      summary: Altera a senha de um usuário específico
      tags:
      - Usuario
servers:
- description: Servidor local de desenvolvimento
  url: http://localhost:8080/api/v1
tags:
- description: Operações para gerenciar os tipos de usuário
  name: TipoUsuario
- description: CRUD de usuários
  name: Usuario
- description: CRUD de restaurantes
  name: Restaurante
- description: Gerenciamento de itens de cardápio
  name: ItemCardapio
components:
  parameters:
    IdParam:
      description: Identificador único do recurso
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    ItemCardapioResponse:
      description: Item do cardápio.
      properties:
        caminhoFoto:
          type: string
        consumoLocal:
          type: boolean
        descricao:
          type: string
        id:
          type: string
          format: uuid
        idRestaurante:
          type: string
          format: uuid
        nome:
          type: string
        preco:
          type: number
      type: object
    ItemCardapioRequest:
      properties:
        caminhoFoto:
          type: string
        consumoLocal:
          type: boolean
        descricao:
          type: string
        nome:
          type: string
        preco:
          type: number
      required:
      - nome
      - preco
      type: object
    RestauranteResponse:
      description: Dados do restaurante.
      properties:
        endereco:
          type: string
        horarioFuncionamento:
          type: string
        id:
          type: string
          format: uuid
        idDono:
          type: string
          format: uuid
        nome:
          type: string
        tipoCozinha:
          type: string
      type: object
    RestauranteRequest:
      properties:
        endereco:
          type: string
        horarioFuncionamento:
          type: string
        idDono:
          type: string
          format: uuid
        nome:
          type: string
        tipoCozinha:
          type: string
      required:
      - nome
      - endereco
      - idDono
      type: object
    AtualizarRestauranteRequest:
      properties:
        endereco:
          type: string
        horarioFuncionamento:
          type: string
        idDono:
          type: string
          format: uuid
        nome:
          type: string
        tipoCozinha:
          type: string
      required:
      - nome
      - endereco
      - idDono
      type: object
    TipoUsuarioResponse:
      description: Representa um tipo de usuário no sistema.
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        categoriaUsuario:
          type: string
      required:
        - categoriaUsuario
        - nome
      type: object
    TipoUsuarioRequest:
      properties:
        nome:
          type: string
        categoriaUsuario:
          type: string
          description: Tipo de usuário
          enum:
            - CLIENTE
            - DONO_RESTAURANTE
            - ADMIN
      required:
        - categoriaUsuario
        - nome
      type: object
    UsuarioRequest:
      type: object
      properties:
        nome:
          type: string
          example: John Doe
        email:
          type: string
          example: johndoe@example.com
        login:
          type: string
          example: johndoe
        senha:
          type: string
          example: secret123
        tipo:
          type: string
          format: uuid
        endereco:
          $ref: '#/components/schemas/EnderecoRequest'
      required:
      - nome
      - email
      - login
      - senha
      - tipo
      - endereco
    UsuarioResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        email:
          type: string
        login:
          type: string
        tipo:
          type: object
          properties:
            id:
              type: string
              format: uuid
            nome:
              type: string
            categoriaUsuario:
              type: string
        endereco:
          $ref: '#/components/schemas/EnderecoResponse'
        dataUltimaAlteracao:
          type: string
          format: date-time
      required:
      - id
      - nome
      - email
      - login
      - tipo
      - endereco
      - dataUltimaAlteracao
    EnderecoRequest:
      type: object
      properties:
        rua:
          type: string
          example: Rua das Flores
        numero:
          type: string
          example: 123
        cidade:
          type: string
          example: Barbacena
        cep:
          type: string
          example: 36200-000
        uf:
          type: string
          example: MG
      required:
      - rua
      - numero
      - cidade
      - cep
      - uf
    EnderecoResponse:
      type: object
      properties:
        rua:
          type: string
          example: Rua das Flores
        numero:
          type: string
          example: 123
        cidade:
          type: string
          example: Barbacena
        cep:
          type: string
          example: 36200-000
        uf:
          type: string
          example: MG
    AlterarSenhaRequest:
      type: object
      properties:
        senhaAtual:
          type: string
          example: secret123
        novaSenha:
          type: string
          example: novaSenha456
      required:
      - senhaAtual
      - novaSenha
    AtualizarUsuarioRequest:
      type: object
      properties:
        nome:
          type: string
        email:
          type: string
          format: email
        login:
          type: string
        tipo:
          type: string
          format: uuid
        endereco:
          $ref: '#/components/schemas/EnderecoRequest'